Refactor data code out to api & tweaks
